+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 7.9.3
+
+Bug fix for validating URLs to allow localhost URLs.
+
+Update to Android Gradle Plugin 8.1.0
+
 ## 7.9.2
 
 Add configuration `Config.useCacheDir` for Android and improved temp file logic. By default (`Config.whenAble`) the downloader will now use the application's `cacheDir` when the size of the file to download is less than half of the `cacheQuotaBytes` given to the app by Android, and use `filesDir` otherwise. If you find that downloads do not complete (or cannot be resumed when paused) this indicates the OS is removing the temp file from the `cacheDir` due to low memory conditions. In that situation, consider using `Config.never` to force the use of `filesDir`, but make sure to clean up remnant temp files in `filesDir`, as the OS does not do that for you.